files splitted create separate files for common utility vs prototype extensions
