+++ html/options.html
@@ -9,6 +9,7 @@
 	<link href="../css/editor.min.css" rel="stylesheet">
 
 	<script src="../js/editor.min.js"></script>
+	<script src="../js/protoExtenders.js"></script>
 	<script src="../js/options.js"></script>
 	<script src="../js/detector.js"></script>
 	<meta name="viewport" content="width=device-width, initial-scale=1">
