+++ webpack.config.js
@@ -8,6 +8,7 @@ module.exports = {
         background: `${__dirname}/js/background.js`,
         detector: `${__dirname}/js/detector.js`,
         options: `${__dirname}/js/options.js`,
+        protoExtenders: `${__dirname}/js/protoExtenders.js`,
     },
     output: {
         filename: "[name].js",
